--- app/build.gradle
@@ -1,50 +1,0 @@
-/*
- * This file was generated by the Gradle 'init' task.
- *
- * This generated file contains a sample Kotlin application project to get you started.
- * For more details take a look at the 'Building Java & JVM projects' chapter in the Gradle
- * User Manual available at https://docs.gradle.org/7.1.1/userguide/building_java_projects.html
- */
-
-plugins {
-    // Apply the org.jetbrains.kotlin.jvm Plugin to add support for Kotlin.
-    id 'org.jetbrains.kotlin.jvm' version '1.4.31'
-
-    // Apply the application plugin to add support for building a CLI application in Java.
-    id 'application'
-}
-
-repositories {
-    // Use Maven Central for resolving dependencies.
-    mavenCentral()
-    maven { url "https://maven.pkg.jetbrains.space/public/p/kotlinx-html/maven" }
-}
-
-dependencies {
-    // Align versions of all Kotlin components
-    implementation platform('org.jetbrains.kotlin:kotlin-bom')
-
-    // Use the Kotlin JDK 8 standard library.
-    implementation 'org.jetbrains.kotlin:kotlin-stdlib-jdk8'
-
-    // This dependency is used by the application.
-    implementation 'com.google.guava:guava:30.1-jre'
-
-    // Use the Kotlin test library.
-    testImplementation 'org.jetbrains.kotlin:kotlin-test'
-
-    // Use the Kotlin JUnit integration.
-    testImplementation 'org.jetbrains.kotlin:kotlin-test-junit'
-
-    // Markdown parser
-    implementation 'org.jetbrains:markdown-jvm:0.2.3'
-
-    // include for JVM target
-    implementation("org.jetbrains.kotlinx:kotlinx-html-jvm:0.7.3")
-    implementation("org.jetbrains.kotlin-wrappers:kotlin-css-jvm:1.0.0-pre.254-kotlin-1.5.31")
-}
-
-application {
-    // Define the main class for the application.
-    mainClass = 'Hacktoberfest2021.AppKt'
-}
