--- app/src/test/kotlin/Hacktoberfest2021/AppTest.kt
@@ -1,14 +1,0 @@
-/*
- * This Kotlin source file was generated by the Gradle 'init' task.
- */
-package Hacktoberfest2021
-
-import kotlin.test.Test
-import kotlin.test.assertNotNull
-
-class AppTest {
-    @Test fun appHasAGreeting() {
-        val classUnderTest = App()
-        assertNotNull(classUnderTest.greeting, "app should have a greeting")
-    }
-}
