+++ explore/issues/android/index.html
@@ -1,0 +1,2 @@
+<!DOCTYPE html>
+<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>Android</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/yogeshpaliyal/ProGuard-Directory/issues/1" target="_blank">Add More Libraries</a></td><td><a href="https://github.com/yogeshpaliyal/ProGuard-Director" target="_blank">ProGuard Directory</a></td><td><em>Android</em></td></tr><tr><td>2</td><td><a href="https://github.com/niharika2810/android-development-best-practices/issues/5" target="_blank">Add more Android developments best practices to help the world</a></td><td><a href="https://github.com/niharika2810/android-development-best-practices" target="_blank">Android Development Best Practices</a></td><td><em>Android</em></td></tr><tr><td>3</td><td><a href="https://github.com/niharika2810/android-interview-questions/issues/4" target="_blank">Add more Interview questions based on your experience</a></td><td><a href="https://github.com/niharika2810/android-interview-questions" target="_blank">Android Interview Questions</a></td><td><em>Android</em></td></tr><tr><td>4</td><td><a href="https://github.com/yogeshpaliyal/Motion-Layout-Samples/issues/1" target="_blank">Add more samples</a></td><td><a href="https://github.com/yogeshpaliyal/Motion-Layout-Samples" target="_blank">Motion-Layout-Samples</a></td><td><em>Android</em></td></tr><tr><td>5</td><td><a href="https://github.com/yogeshpaliyal/Android-Universal-Recycler-View-Adapter/issues/13" target="_blank">Add More Examples</a></td><td><a href="https://github.com/yogeshpaliyal/Android-Universal-Recycler-View-Adapter" target="_blank">Android-Universal-Recycler-View-Adapter</a></td><td><em>Android</em><em>Kotlin</em></td></tr><tr><td>6</td><td><a href="https://github.com/yogeshpaliyal/Android-Universal-Recycler-View-Adapter/issues/12" target="_blank">Readme is very cluttered</a></td><td><a href="https://github.com/yogeshpaliyal/Android-Universal-Recycler-View-Adapter" target="_blank">Android-Universal-Recycler-View-Adapter</a></td><td><em>Android</em><em>Markdown</em></td></tr></table></body></html>
