+++ explore/issues/any language/index.html
@@ -1,2 +1,2 @@
 <!DOCTYPE html>
-<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>Any language</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/vinitshahdeo/HBD/issues/1" target="_blank">Coder's Way of Wishing Happy Birthday</a></td><td><a href="https://github.com/vinitshahdeo/HBD/" target="_blank">HBD</a></td><td><em>Any Language</em></td></tr></table></body></html>
+<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>Any language</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/vinitshahdeo/HBD/issues/1" target="_blank">Coder's Way of Wishing Happy Birthday</a></td><td><a href="https://github.com/vinitshahdeo/HBD/" target="_blank">HBD</a></td><td><em>Any Language</em></td></tr><tr><td>2</td><td><a href="https://github.com/TerexitariusStomp/treeauto/issues/1" target="_blank">Initial Coding Help</a></td><td><a href="https://github.com/TerexitariusStomp/treeauto" target="_blank">Tree Auto Planting</a></td><td><em>Any Language</em></td></tr></table></body></html>
