+++ explore/issues/bootstrap/index.html
@@ -1,2 +1,2 @@
 <!DOCTYPE html>
-<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>Bootstrap</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/bacloud14/Classified-ads-48/issues/6" target="_blank">Correct Bootstrap layout and forms</a></td><td><a href="https://github.com/bacloud14/Classified-ads-48/" target="_blank">Classified-ads-48</a></td><td><em>HTML</em><em>JavaScript</em><em>Bootstrap</em><em>Node.js</em></td></tr><tr><td>2</td><td><a href="https://github.com/bacloud14/Classified-ads-48/issues/8" target="_blank">Improve Google Lighthouse score and other SEO audits</a></td><td><a href="https://github.com/bacloud14/Classified-ads-48/" target="_blank">Classified-ads-48</a></td><td><em>HTML</em><em>JavaScript</em><em>Bootstrap</em><em>Node.js</em></td></tr></table></body></html>
+<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>Bootstrap</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/bacloud14/Classified-ads-48/issues/6" target="_blank">Correct Bootstrap layout and forms</a></td><td><a href="https://github.com/bacloud14/Classified-ads-48/" target="_blank">Classified-ads-48</a></td><td><em>HTML</em><em>JavaScript</em><em>Bootstrap</em><em>Node.js</em></td></tr><tr><td>2</td><td><a href="https://github.com/bacloud14/Classified-ads-48/issues/8" target="_blank">Improve Google Lighthouse score and other SEO audits</a></td><td><a href="https://github.com/bacloud14/Classified-ads-48/" target="_blank">Classified-ads-48</a></td><td><em>HTML</em><em>JavaScript</em><em>Bootstrap</em><em>Node.js</em></td></tr><tr><td>3</td><td><a href="https://github.com/vinitshahdeo/github-stars-feed/issues/12" target="_blank">Create a React App to show GitHub Star's feed in Card View</a></td><td><a href="https://github.com/vinitshahdeo/github-stars-feed" target="_blank">GitHub Stars Feed</a></td><td><em>React</em><em>Bootstrap</em></td></tr></table></body></html>
