+++ explore/issues/c++/index.html
@@ -1,0 +1,2 @@
+<!DOCTYPE html>
+<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>C++</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/Astrodevil/Programming-Basics/issues/4" target="_blank">Add Some C++ files</a></td><td><a href="https://github.com/Astrodevil/Programming-Basics" target="_blank">Programming-Basics </a></td><td><em>C++</em></td></tr><tr><td>2</td><td><a href="https://github.com/prathamesh-borse/Hacktoberfest_2021/issues/1" target="_blank">Add More Projects</a></td><td><a href="https://github.com/prathamesh-borse/Hacktoberfest_2021/" target="_blank">Hacktoberfest2021</a></td><td><em>C++</em><em>Java</em><em>HTML</em><em>Flutter</em></td></tr></table></body></html>
