+++ explore/issues/flutter/index.html
@@ -1,2 +1,2 @@
 <!DOCTYPE html>
-<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>Flutter</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/thisisamank/resizrr/issues/4" target="_blank">Add dark mode to &lt;code&gt;resizrr&lt;/code&gt; app.</a></td><td><a href="https://github.com/thisisamank/resizrr/" target="_blank">Resizrr</a></td><td><em>Dart</em><em>Flutter</em></td></tr></table></body></html>
+<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>Flutter</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/thisisamank/resizrr/issues/4" target="_blank">Add dark mode to &lt;code&gt;resizrr&lt;/code&gt; app.</a></td><td><a href="https://github.com/thisisamank/resizrr/" target="_blank">Resizrr</a></td><td><em>Dart</em><em>Flutter</em></td></tr><tr><td>2</td><td><a href="https://github.com/cankush625/Instagram_Clone/issues/8" target="_blank">Make camera button working</a></td><td><a href="https://github.com/cankush625/Instagram_Clone" target="_blank">Instagram Clone</a></td><td><em>Flutter</em><em>Dart</em></td></tr><tr><td>3</td><td><a href="https://github.com/cankush625/Chat-App/issues/4" target="_blank">Improve Profile Screen</a></td><td><a href="https://github.com/cankush625/Chat-App" target="_blank">Chat-App</a></td><td><em>Flutter</em><em>Dart</em><em>Google Maps API</em><em>Firebase</em></td></tr><tr><td>4</td><td><a href="https://github.com/adityathakurxd/university/issues/24" target="_blank">Fix Sign In Screen</a></td><td><a href="https://github.com/adityathakurxd/university" target="_blank">University</a></td><td><em>Flutter</em><em>Dart</em><em>Firebase</em></td></tr><tr><td>5</td><td><a href="https://github.com/prathamesh-borse/Hacktoberfest_2021/issues/1" target="_blank">Add More Projects</a></td><td><a href="https://github.com/prathamesh-borse/Hacktoberfest_2021/" target="_blank">Hacktoberfest2021</a></td><td><em>C++</em><em>Java</em><em>HTML</em><em>Flutter</em></td></tr></table></body></html>
