+++ explore/issues/java/index.html
@@ -1,2 +1,2 @@
 <!DOCTYPE html>
-<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>Java</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/utkarsh1504/DSA-Java/issues/1" target="_blank">new lesson: Introduction to Arrays</a></td><td><a href="https://github.com/utkarsh1504/DSA-Java/" target="_blank">DSA-Java</a></td><td><em>Java</em><em>Markdown</em><em>Images</em></td></tr></table></body></html>
+<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>Java</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/utkarsh1504/DSA-Java/issues/1" target="_blank">new lesson: Introduction to Arrays</a></td><td><a href="https://github.com/utkarsh1504/DSA-Java/" target="_blank">DSA-Java</a></td><td><em>Java</em><em>Markdown</em><em>Images</em></td></tr><tr><td>2</td><td><a href="https://github.com/strongbox/strongbox/issues/1730" target="_blank">Centralize Documentation Links</a></td><td><a href="https://github.com/strongbox/strongbox" target="_blank">strongbox</a></td><td><em>Java</em><em>Documentation</em></td></tr><tr><td>3</td><td><a href="https://github.com/appsmithorg/appsmith/issues?q=is%3Aopen+is%3Aissue+bug+label%3A%22Good+First+Issue%22" target="_blank">Good first issues</a></td><td><a href="https://github.com/appsmithorg/appsmith" target="_blank">Appsmith</a></td><td><em>JavaScript</em><em>Java</em></td></tr><tr><td>4</td><td><a href="https://github.com/prathamesh-borse/Hacktoberfest_2021/issues/1" target="_blank">Add More Projects</a></td><td><a href="https://github.com/prathamesh-borse/Hacktoberfest_2021/" target="_blank">Hacktoberfest2021</a></td><td><em>C++</em><em>Java</em><em>HTML</em><em>Flutter</em></td></tr></table></body></html>
