+++ explore/issues/js/index.html
@@ -1,2 +1,2 @@
 <!DOCTYPE html>
-<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>Js</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/vinitshahdeo/SimpleBio/issues/4" target="_blank">Add more templates</a></td><td><a href="https://github.com/vinitshahdeo/SimpleBio" target="_blank">SimpleBio</a></td><td><em>HTML</em><em>CSS</em><em>JS</em></td></tr><tr><td>2</td><td><a href="https://github.com/vinitshahdeo/online-debate-system/issues/3" target="_blank">UI Improvements</a></td><td><a href="https://github.com/vinitshahdeo/online-debate-system" target="_blank">online-debate-system</a></td><td><em>HTML</em><em>CSS</em><em>JS</em></td></tr><tr><td>3</td><td><a href="https://github.com/vinitshahdeo/water-monitoring-system/issues/2" target="_blank">UI Enhancements</a></td><td><a href="https://github.com/vinitshahdeo/water-monitoring-system/" target="_blank">Water Monitoring System</a></td><td><em>HTML</em><em>CSS</em><em>JS</em></td></tr><tr><td>4</td><td><a href="https://github.com/vinitshahdeo/water-monitoring-system/issues/5" target="_blank">Admin Dashboard</a></td><td><a href="https://github.com/vinitshahdeo/water-monitoring-system/" target="_blank">Water Monitoring System</a></td><td><em>HTML</em><em>CSS</em><em>JS</em></td></tr></table></body></html>
+<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>Js</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/vinitshahdeo/SimpleBio/issues/4" target="_blank">Add more templates</a></td><td><a href="https://github.com/vinitshahdeo/SimpleBio" target="_blank">SimpleBio</a></td><td><em>HTML</em><em>CSS</em><em>JS</em></td></tr><tr><td>2</td><td><a href="https://github.com/vinitshahdeo/online-debate-system/issues/3" target="_blank">UI Improvements</a></td><td><a href="https://github.com/vinitshahdeo/online-debate-system" target="_blank">online-debate-system</a></td><td><em>HTML</em><em>CSS</em><em>JS</em></td></tr><tr><td>3</td><td><a href="https://github.com/vinitshahdeo/water-monitoring-system/issues/2" target="_blank">UI Enhancements</a></td><td><a href="https://github.com/vinitshahdeo/water-monitoring-system/" target="_blank">Water Monitoring System</a></td><td><em>HTML</em><em>CSS</em><em>JS</em></td></tr><tr><td>4</td><td><a href="https://github.com/vinitshahdeo/water-monitoring-system/issues/5" target="_blank">Admin Dashboard</a></td><td><a href="https://github.com/vinitshahdeo/water-monitoring-system/" target="_blank">Water Monitoring System</a></td><td><em>HTML</em><em>CSS</em><em>JS</em></td></tr><tr><td>5</td><td><a href="https://github.com/saloni-15/Heel2Toe/issues/1" target="_blank">Improve Footer Section</a></td><td><a href="https://github.com/saloni-15/Heel2Toe" target="_blank">Heel2Toe</a></td><td><em>HTML</em><em>CSS</em><em>JS</em></td></tr><tr><td>6</td><td><a href="https://github.com/saloni-15/Heel2Toe/issues/31" target="_blank">Add a navbar</a></td><td><a href="https://github.com/saloni-15/Heel2Toe" target="_blank">Heel2Toe</a></td><td><em>HTML</em><em>CSS</em><em>JS</em></td></tr><tr><td>7</td><td><a href="https://github.com/TryGhost/Ghost/issues/13216" target="_blank">Unable to import Members</a></td><td><a href="https://github.com/TryGhost/Ghost" target="_blank">Ghost</a></td><td><em>HTML</em><em>CSS</em><em>JS</em></td></tr><tr><td>8</td><td><a href="https://github.com/Opentek-Org/opentek/issues/64" target="_blank">Redesign and Add New Cards</a></td><td><a href="https://github.com/Opentek-Org/opentek" target="_blank">Opentek Web</a></td><td><em>HTML</em><em>CSS</em><em>JS</em></td></tr><tr><td>9</td><td><a href="https://github.com/Opentek-Org/opentek/issues/69" target="_blank">Add more OPEN SOURCE programs and events </a></td><td><a href="https://github.com/Opentek-Org/opentek" target="_blank">Opentek Web</a></td><td><em>HTML</em><em>CSS</em><em>JS</em></td></tr></table></body></html>
