+++ explore/issues/kmm/index.html
@@ -1,0 +1,2 @@
+<!DOCTYPE html>
+<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>Kmm</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/shmehdi01/quote-app-kmm/issues" target="_blank">Improve Login Screen (Compose)</a></td><td><a href="https://github.com/shmehdi01/quote-app-kmm" target="_blank">Quote APP(KMM)</a></td><td><em>Kotlin</em><em>KMM</em></td></tr></table></body></html>
