+++ explore/issues/kotlin/index.html
@@ -1,2 +1,2 @@
 <!DOCTYPE html>
-<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>Kotlin</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/KunalRaghav/Crossfire/issues/1" target="_blank">Add more questions</a></td><td><a href="https://github.com/KunalRaghav/Crossfire" target="_blank">CrossFire ❌🔥</a></td><td><em>Kotlin</em></td></tr></table></body></html>
+<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>Kotlin</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/KunalRaghav/Crossfire/issues/1" target="_blank">Add more questions</a></td><td><a href="https://github.com/KunalRaghav/Crossfire" target="_blank">CrossFire ❌🔥</a></td><td><em>Kotlin</em></td></tr><tr><td>2</td><td><a href="https://github.com/KunalRaghav/Crossfire/issues/2" target="_blank">Use android viewBindings</a></td><td><a href="https://github.com/KunalRaghav/Crossfire" target="_blank">CrossFire ❌🔥</a></td><td><em>Kotlin</em></td></tr><tr><td>3</td><td><a href="https://github.com/niharika2810/MovieTray/issues/2" target="_blank">Try to find issues with the app and propose a solution.</a></td><td><a href="https://github.com/niharika2810/MovieTray" target="_blank">MovieTray</a></td><td><em>Kotlin</em></td></tr><tr><td>4</td><td><a href="https://github.com/shmehdi01/quote_api_ktor/issues" target="_blank">Add Profile Api For User</a></td><td><a href="https://github.com/shmehdi01/quote_api_ktor" target="_blank">Ktor Quote API</a></td><td><em>Kotlin</em><em>Ktor</em></td></tr><tr><td>5</td><td><a href="https://github.com/shmehdi01/quote-app-kmm/issues" target="_blank">Improve Login Screen (Compose)</a></td><td><a href="https://github.com/shmehdi01/quote-app-kmm" target="_blank">Quote APP(KMM)</a></td><td><em>Kotlin</em><em>KMM</em></td></tr><tr><td>6</td><td><a href="https://github.com/yogeshpaliyal/Android-Universal-Recycler-View-Adapter/issues/13" target="_blank">Add More Examples</a></td><td><a href="https://github.com/yogeshpaliyal/Android-Universal-Recycler-View-Adapter" target="_blank">Android-Universal-Recycler-View-Adapter</a></td><td><em>Android</em><em>Kotlin</em></td></tr></table></body></html>
