+++ explore/issues/markdown/index.html
@@ -1,2 +1,2 @@
 <!DOCTYPE html>
-<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>Markdown</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/utkarsh1504/DSA-Java/issues/1" target="_blank">new lesson: Introduction to Arrays</a></td><td><a href="https://github.com/utkarsh1504/DSA-Java/" target="_blank">DSA-Java</a></td><td><em>Java</em><em>Markdown</em><em>Images</em></td></tr><tr><td>2</td><td><a href="https://github.com/vinitshahdeo/water-monitoring-system/issues/3" target="_blank">Project Documentation</a></td><td><a href="https://github.com/vinitshahdeo/water-monitoring-system/" target="_blank">Water Monitoring System</a></td><td><em>Markdown</em></td></tr></table></body></html>
+<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>Markdown</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/utkarsh1504/DSA-Java/issues/1" target="_blank">new lesson: Introduction to Arrays</a></td><td><a href="https://github.com/utkarsh1504/DSA-Java/" target="_blank">DSA-Java</a></td><td><em>Java</em><em>Markdown</em><em>Images</em></td></tr><tr><td>2</td><td><a href="https://github.com/vinitshahdeo/water-monitoring-system/issues/3" target="_blank">Project Documentation</a></td><td><a href="https://github.com/vinitshahdeo/water-monitoring-system/" target="_blank">Water Monitoring System</a></td><td><em>Markdown</em></td></tr><tr><td>3</td><td><a href="https://github.com/yogeshpaliyal/Android-Universal-Recycler-View-Adapter/issues/12" target="_blank">Readme is very cluttered</a></td><td><a href="https://github.com/yogeshpaliyal/Android-Universal-Recycler-View-Adapter" target="_blank">Android-Universal-Recycler-View-Adapter</a></td><td><em>Android</em><em>Markdown</em></td></tr><tr><td>4</td><td><a href="https://github.com/gauranshkumar/chithi/issues/2" target="_blank">Update the Readme for the project</a></td><td><a href="https://github.com/gauranshkumar/chithi" target="_blank">Chitthi</a></td><td><em>Markdown</em><em>Documentaion</em></td></tr><tr><td>5</td><td><a href="https://github.com/gauranshkumar/chithi/issues/5" target="_blank">Add New feature for Markdown Shortcut Supported Editor</a></td><td><a href="https://github.com/gauranshkumar/chithi" target="_blank">Chitthi</a></td><td><em>Python 3</em><em>Streamlit Library</em><em>Javascript</em><em>HTML/CSS</em><em>Markdown</em></td></tr><tr><td>6</td><td><a href="https://github.com/chandrikadeb7/Hacktoberfest_2021/issues/206" target="_blank">Project Readme Ideas (Open to Discussion)</a></td><td><a href="https://github.com/chandrikadeb7/Hacktoberfest_2021" target="_blank">Hacktoberfest_2021</a></td><td><em>Markdown</em></td></tr></table></body></html>
