+++ explore/issues/nodejs/index.html
@@ -1,2 +1,2 @@
 <!DOCTYPE html>
-<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>Nodejs</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/vinitshahdeo/Hackathon-Timer/issues/1" target="_blank">Use GitHub webhooks</a></td><td><a href="https://github.com/vinitshahdeo/Hackathon-Timer" target="_blank">Hackathon-Timer</a></td><td><em>NodeJS</em></td></tr><tr><td>2</td><td><a href="https://github.com/vinitshahdeo/Quotter/issues/6" target="_blank">Auto follower (Follow for Follow)</a></td><td><a href="https://github.com/vinitshahdeo/Quotter/" target="_blank">Quotter</a></td><td><em>NodeJS</em></td></tr><tr><td>3</td><td><a href="https://github.com/vinitshahdeo/Quotter/issues/5" target="_blank">Adding a feature to DM the quotes to all the followers</a></td><td><a href="https://github.com/vinitshahdeo/Quotter/" target="_blank">Quotter</a></td><td><em>NodeJS</em></td></tr><tr><td>4</td><td><a href="https://github.com/vinitshahdeo/Quotter/issues/12" target="_blank">.gitignore is missing</a></td><td><a href="https://github.com/vinitshahdeo/Quotter/" target="_blank">Quotter</a></td><td><em>NodeJS</em></td></tr><tr><td>5</td><td><a href="https://github.com/felixfaisal/attack-on-web/issues/31" target="_blank">Fix Linting</a></td><td><a href="https://github.com/felixfaisal/attack-on-web" target="_blank">Attack on Web</a></td><td><em>JavaScript</em><em>NodeJs</em></td></tr></table></body></html>
+<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>Nodejs</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/vinitshahdeo/Hackathon-Timer/issues/1" target="_blank">Use GitHub webhooks</a></td><td><a href="https://github.com/vinitshahdeo/Hackathon-Timer" target="_blank">Hackathon-Timer</a></td><td><em>NodeJS</em></td></tr><tr><td>2</td><td><a href="https://github.com/vinitshahdeo/Quotter/issues/6" target="_blank">Auto follower (Follow for Follow)</a></td><td><a href="https://github.com/vinitshahdeo/Quotter/" target="_blank">Quotter</a></td><td><em>NodeJS</em></td></tr><tr><td>3</td><td><a href="https://github.com/vinitshahdeo/Quotter/issues/5" target="_blank">Adding a feature to DM the quotes to all the followers</a></td><td><a href="https://github.com/vinitshahdeo/Quotter/" target="_blank">Quotter</a></td><td><em>NodeJS</em></td></tr><tr><td>4</td><td><a href="https://github.com/vinitshahdeo/Quotter/issues/12" target="_blank">.gitignore is missing</a></td><td><a href="https://github.com/vinitshahdeo/Quotter/" target="_blank">Quotter</a></td><td><em>NodeJS</em></td></tr><tr><td>5</td><td><a href="https://github.com/felixfaisal/attack-on-web/issues/31" target="_blank">Fix Linting</a></td><td><a href="https://github.com/felixfaisal/attack-on-web" target="_blank">Attack on Web</a></td><td><em>JavaScript</em><em>NodeJs</em></td></tr><tr><td>6</td><td><a href="https://github.com/Opentek-Org/opentek/issues/18" target="_blank">Add Hackathon Page using Web Scraping</a></td><td><a href="https://github.com/Opentek-Org/opentek" target="_blank">Opentek Web</a></td><td><em>HTML</em><em>CSS</em><em>NodeJS</em></td></tr></table></body></html>
