+++ explore/issues/python 3/index.html
@@ -1,0 +1,2 @@
+<!DOCTYPE html>
+<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>Python 3</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/gauranshkumar/chithi/issues/4" target="_blank">Refactoring the UI Code</a></td><td><a href="https://github.com/gauranshkumar/chithi" target="_blank">Chitthi</a></td><td><em>Python 3</em><em>Streamlit Library</em></td></tr><tr><td>2</td><td><a href="https://github.com/gauranshkumar/chithi/issues/6" target="_blank">Solidifiaction of the current App</a></td><td><a href="https://github.com/gauranshkumar/chithi" target="_blank">Chitthi</a></td><td><em>Python 3</em><em>Streamlit Library</em></td></tr><tr><td>3</td><td><a href="https://github.com/gauranshkumar/chithi/issues/5" target="_blank">Add New feature for Markdown Shortcut Supported Editor</a></td><td><a href="https://github.com/gauranshkumar/chithi" target="_blank">Chitthi</a></td><td><em>Python 3</em><em>Streamlit Library</em><em>Javascript</em><em>HTML/CSS</em><em>Markdown</em></td></tr></table></body></html>
