+++ explore/issues/python/index.html
@@ -1,2 +1,2 @@
 <!DOCTYPE html>
-<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>Python</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/vinitshahdeo/jobtweets/issues/3" target="_blank">Taking query(#hashtags) as input from user</a></td><td><a href="https://github.com/vinitshahdeo/jobtweets/" target="_blank">jobtweets</a></td><td><em>Python</em></td></tr><tr><td>2</td><td><a href="https://github.com/vinitshahdeo/jobtweets/issues/28" target="_blank">[Pie Chart] Graphical representation of the results</a></td><td><a href="https://github.com/vinitshahdeo/jobtweets/" target="_blank">jobtweets</a></td><td><em>Python</em></td></tr><tr><td>3</td><td><a href="https://github.com/dry-python/classes/issues/307" target="_blank">Add &lt;code&gt;Any&lt;/code&gt; alias to &lt;code&gt;object&lt;/code&gt;</a></td><td><a href="https://github.com/dry-python/classes" target="_blank">dry-python/classes</a></td><td><em>Python</em></td></tr></table></body></html>
+<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>Python</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/vinitshahdeo/jobtweets/issues/3" target="_blank">Taking query(#hashtags) as input from user</a></td><td><a href="https://github.com/vinitshahdeo/jobtweets/" target="_blank">jobtweets</a></td><td><em>Python</em></td></tr><tr><td>2</td><td><a href="https://github.com/vinitshahdeo/jobtweets/issues/28" target="_blank">[Pie Chart] Graphical representation of the results</a></td><td><a href="https://github.com/vinitshahdeo/jobtweets/" target="_blank">jobtweets</a></td><td><em>Python</em></td></tr><tr><td>3</td><td><a href="https://github.com/dry-python/classes/issues/307" target="_blank">Add &lt;code&gt;Any&lt;/code&gt; alias to &lt;code&gt;object&lt;/code&gt;</a></td><td><a href="https://github.com/dry-python/classes" target="_blank">dry-python/classes</a></td><td><em>Python</em></td></tr><tr><td>4</td><td><a href="https://github.com/scrapy/scrapy/issues/5135" target="_blank">Remove UrlLengthMiddleware from default enabled middlewares</a></td><td><a href="https://github.com/scrapy/scrapy" target="_blank">Scrapy</a></td><td><em>Python</em></td></tr><tr><td>5</td><td><a href="https://github.com/krassowski/jupyterlab-lsp/issues/362" target="_blank">Make hover tooltip show up automatically after specified delay</a></td><td><a href="https://github.com/krassowski/jupyterlab-lsp" target="_blank">jupyterlab-lsp</a></td><td><em>Python</em><em>TypeScript</em><em>Python</em><em>Robot-Framework</em><em>JavaScript</em></td></tr><tr><td>6</td><td><a href="https://github.com/TeamUltroid/Ultroid/issues/175" target="_blank">Help to Complete Translation work.</a></td><td><a href="https://github.com/TeamUltroid/Ultroid" target="_blank">Ultroid Userbot</a></td><td><em>Python</em></td></tr><tr><td>7</td><td><a href="https://github.com/cankush625/PlanMyTrip/issues/1" target="_blank">Refactor the News model</a></td><td><a href="https://github.com/cankush625/PlanMyTrip" target="_blank">PlanMyTrip</a></td><td><em>Django</em><em>Python</em></td></tr><tr><td>8</td><td><a href="https://github.com/Ccode-lang/echo-lang/issues/1" target="_blank">needs more commands</a></td><td><a href="https://github.com/Ccode-lang/echo-lang" target="_blank">Echo language</a></td><td><em>Python</em></td></tr></table></body></html>
