+++ explore/issues/python3/index.html
@@ -1,0 +1,2 @@
+<!DOCTYPE html>
+<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>Python3</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/Abhinav-26/Machine-Learning-Minor-Projects/issues" target="_blank">[ENHANCEMENT] Organise the projects under appropriate topics</a></td><td><a href="https://github.com/Abhinav-26/Machine-Learning-Minor-Projects" target="_blank">Machine Learning Minor Projects</a></td><td><em>Python3</em></td></tr></table></body></html>
