+++ explore/issues/react/index.html
@@ -1,2 +1,2 @@
 <!DOCTYPE html>
-<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>React</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/vinitshahdeo/MiniYouTube/issues/1" target="_blank">Add ESLint</a></td><td><a href="https://github.com/vinitshahdeo/MiniYouTube" target="_blank">MiniYouTube</a></td><td><em>React</em></td></tr><tr><td>2</td><td><a href="https://github.com/vinitshahdeo/myFaculty/issues/2" target="_blank">UI improvisation</a></td><td><a href="https://github.com/vinitshahdeo/myFaculty/" target="_blank">myFaculty</a></td><td><em>React</em></td></tr></table></body></html>
+<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>React</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/vinitshahdeo/MiniYouTube/issues/1" target="_blank">Add ESLint</a></td><td><a href="https://github.com/vinitshahdeo/MiniYouTube" target="_blank">MiniYouTube</a></td><td><em>React</em></td></tr><tr><td>2</td><td><a href="https://github.com/vinitshahdeo/myFaculty/issues/2" target="_blank">UI improvisation</a></td><td><a href="https://github.com/vinitshahdeo/myFaculty/" target="_blank">myFaculty</a></td><td><em>React</em></td></tr><tr><td>3</td><td><a href="https://github.com/vinitshahdeo/Corona-India/issues/37" target="_blank">News API is not working</a></td><td><a href="https://github.com/vinitshahdeo/Corona-India" target="_blank">COVID-19 Tracker 🇮🇳</a></td><td><em>React</em></td></tr><tr><td>4</td><td><a href="https://github.com/vinitshahdeo/github-stars-feed/issues/12" target="_blank">Create a React App to show GitHub Star's feed in Card View</a></td><td><a href="https://github.com/vinitshahdeo/github-stars-feed" target="_blank">GitHub Stars Feed</a></td><td><em>React</em><em>Bootstrap</em></td></tr><tr><td>5</td><td><a href="https://github.com/aman34503/Go-Airbnb/issues/1" target="_blank">Create More Functionality_and Added Authorisation</a></td><td><a href="https://github.com/aman34503/Go-Airbnb" target="_blank">Go-Airbnb</a></td><td><em>React</em><em>MaterialUI</em></td></tr><tr><td>6</td><td><a href="https://github.com/vinitshahdeo/hacktoberfest-issue-hunt/issues/1" target="_blank">Create a React App to filter Hacktoberfest issues on the basis of languages and labels</a></td><td><a href="https://github.com/vinitshahdeo/hacktoberfest-issue-hunt" target="_blank">hacktoberfest-issue-hunt</a></td><td><em>React</em></td></tr></table></body></html>
