+++ explore/issues/reactjs/index.html
@@ -1,2 +1,2 @@
 <!DOCTYPE html>
-<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>Reactjs</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/rishipurwar1/coding-space/issues/28" target="_blank">Create a Join Discord Community button</a></td><td><a href="https://github.com/rishipurwar1/coding-space" target="_blank">CODINGSPACE</a></td><td><em>REACTJS</em><em>TAILWINDCSS</em><em>FIREBASE</em></td></tr><tr><td>2</td><td><a href="https://github.com/rishipurwar1/coding-space/issues/24" target="_blank">Fix challenge details page blank screen</a></td><td><a href="https://github.com/rishipurwar1/coding-space" target="_blank">CODINGSPACE</a></td><td><em>REACTJS</em><em>TAILWINDCSS</em><em>FIREBASE</em></td></tr><tr><td>3</td><td><a href="https://github.com/rishipurwar1/coding-space/issues/26" target="_blank">Card and button lack hover effect</a></td><td><a href="https://github.com/rishipurwar1/coding-space" target="_blank">CODINGSPACE</a></td><td><em>REACTJS</em><em>TAILWINDCSS</em><em>FIREBASE</em></td></tr></table></body></html>
+<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>Reactjs</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/rishipurwar1/coding-space/issues/28" target="_blank">Create a Join Discord Community button</a></td><td><a href="https://github.com/rishipurwar1/coding-space" target="_blank">CODINGSPACE</a></td><td><em>REACTJS</em><em>TAILWINDCSS</em><em>FIREBASE</em></td></tr><tr><td>2</td><td><a href="https://github.com/rishipurwar1/coding-space/issues/24" target="_blank">Fix challenge details page blank screen</a></td><td><a href="https://github.com/rishipurwar1/coding-space" target="_blank">CODINGSPACE</a></td><td><em>REACTJS</em><em>TAILWINDCSS</em><em>FIREBASE</em></td></tr><tr><td>3</td><td><a href="https://github.com/rishipurwar1/coding-space/issues/26" target="_blank">Card and button lack hover effect</a></td><td><a href="https://github.com/rishipurwar1/coding-space" target="_blank">CODINGSPACE</a></td><td><em>REACTJS</em><em>TAILWINDCSS</em><em>FIREBASE</em></td></tr><tr><td>4</td><td><a href="https://github.com/Sin-Sumit/Accordion/issues" target="_blank">Make it Dynamic</a></td><td><a href="https://github.com/Sin-Sumit/Accordion" target="_blank">Accordion</a></td><td><em>ReactJs</em></td></tr></table></body></html>
