+++ explore/issues/robot-framework/index.html
@@ -1,0 +1,2 @@
+<!DOCTYPE html>
+<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>Robot-framework</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/krassowski/jupyterlab-lsp/issues/362" target="_blank">Make hover tooltip show up automatically after specified delay</a></td><td><a href="https://github.com/krassowski/jupyterlab-lsp" target="_blank">jupyterlab-lsp</a></td><td><em>Python</em><em>TypeScript</em><em>Python</em><em>Robot-Framework</em><em>JavaScript</em></td></tr></table></body></html>
