+++ explore/issues/scss/index.html
@@ -1,0 +1,2 @@
+<!DOCTYPE html>
+<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>Scss</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/chatwoot/chatwoot/issues/3081" target="_blank">List is not rendering properly in the dashboard UI</a></td><td><a href="https://github.com/chatwoot/chatwoot" target="_blank">Chatwoot</a></td><td><em>HTML</em><em>SCSS</em><em>JavaScript</em><em>Ruby</em><em>Vue.Js</em></td></tr></table></body></html>
