+++ explore/issues/tailwind css/index.html
@@ -1,0 +1,2 @@
+<!DOCTYPE html>
+<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>Tailwind css</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/devstrons/heloworld.xyz/issues/68" target="_blank">[BUG] Main webpage throws warnings</a></td><td><a href="https://github.com/devstrons/heloworld.xyz" target="_blank">heloworld.xyz 👋</a></td><td><em>Javascript</em><em>Tailwind CSS</em><em>website</em></td></tr><tr><td>2</td><td><a href="https://github.com/devstrons/heloworld.xyz/issues/69" target="_blank">[BUG] Refresh animation hidden behind &amp;quot;Contribute&amp;quot; button</a></td><td><a href="https://github.com/devstrons/heloworld.xyz" target="_blank">heloworld.xyz 👋</a></td><td><em>Javascript</em><em>Tailwind CSS</em><em>website</em></td></tr></table></body></html>
