+++ explore/issues/typescript/index.html
@@ -1,2 +1,2 @@
 <!DOCTYPE html>
-<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>Typescript</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/Rishabh-malhotraa/caucus/issues/22" target="_blank">Refactoring Code</a></td><td><a href="https://github.com/Rishabh-malhotraa/caucus/" target="_blank">Caucus</a></td><td><em>TypeScript</em></td></tr><tr><td>2</td><td><a href="https://github.com/Rishabh-malhotraa/caucus/issues/8" target="_blank">Adding Zen Mode</a></td><td><a href="https://github.com/Rishabh-malhotraa/caucus/" target="_blank">Caucus</a></td><td><em>TypeScript</em></td></tr></table></body></html>
+<html><head><title>Hacktoberfest 2021 Issues</title><link href="../../.meta/style.css" rel="stylesheet"></head><body><img src="../../.meta/logo.png" class="center"><h1>Typescript</h1><h4><em>This is a generated file from </em><a href="../../ISSUES.md">ISSUES</a><em> kindly don't delete it</em><em>- Automation script by <a href="https://chetangupta.net/about" target="_blank">Chetan Gupta</a></em></h4><table><tr><th>Serial No.</th><th>Issue</th><th>Repository</th><th>Language</th></tr><tr><td>1</td><td><a href="https://github.com/Rishabh-malhotraa/caucus/issues/22" target="_blank">Refactoring Code</a></td><td><a href="https://github.com/Rishabh-malhotraa/caucus/" target="_blank">Caucus</a></td><td><em>TypeScript</em></td></tr><tr><td>2</td><td><a href="https://github.com/Rishabh-malhotraa/caucus/issues/8" target="_blank">Adding Zen Mode</a></td><td><a href="https://github.com/Rishabh-malhotraa/caucus/" target="_blank">Caucus</a></td><td><em>TypeScript</em></td></tr><tr><td>3</td><td><a href="https://github.com/microsoft/TypeScript/issues/45876" target="_blank">Improve Documentation</a></td><td><a href="https://github.com/microsoft/TypeScript" target="_blank">TypeScript</a></td><td><em>JavaScript</em><em>TypeScript</em></td></tr><tr><td>4</td><td><a href="https://github.com/krassowski/jupyterlab-lsp/issues/362" target="_blank">Make hover tooltip show up automatically after specified delay</a></td><td><a href="https://github.com/krassowski/jupyterlab-lsp" target="_blank">jupyterlab-lsp</a></td><td><em>Python</em><em>TypeScript</em><em>Python</em><em>Robot-Framework</em><em>JavaScript</em></td></tr></table></body></html>
